requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-colorfield==0.10.1
django-compressor==4.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.2
django-guardian==2.4.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-mptt==0.14.0
django-quill-editor==0.1.40
django-redis==5.4.0
django-storages==1.14
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.5
packaging==23.2
Pillow==10.0.1
psycopg
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.1
rjsmin==1.2.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
whitenoise==6.5.0
